requirements.txt
boto3
streamlit
langchain
langchain-community
faiss-cpu
python-dotenv
pypdf
